backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
PyPDF2==3.0.1
python-docx==0.8.11
groq==0.9.0
python-socketio==5.8.0
python-engineio==4.7.1
Werkzeug==2.3.7
watchdog==3.0.0
